Move `WeightConfig` into the `weights` directory
Test plan: `yarn test` sufficies for this simple move.
